+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-22 14:12:50
-// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
+// Updated at: 2026-05-22 18:41:52
+// Commit: test(perf): implement widget integration tests for perf
 
 class perfComponentState {
-  final String commitRef = '1320';
-  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
-  final DateTime timestamp = DateTime.parse('2026-05-22T14:12:50+05:00');
+  final String commitRef = '1330';
+  final String lastMessage = 'test(perf): implement widget integration tests for perf';
+  final DateTime timestamp = DateTime.parse('2026-05-22T18:41:52+05:00');
 }