+++ CHANGELOG.md
@@ -1331,3 +1331,4 @@
 - [2026-05-22 17:25] fix(i18n): prevent double tap duplicate action in i18n
 - [2026-05-22 18:30] fix(i18n): resolve race condition during i18n initialization
 - [2026-05-22 18:39] fix(navigation): resolve race condition during navigation initialization
+- [2026-05-22 18:41] test(perf): implement widget integration tests for perf